py/pyproject.toml
[project]
name = "cli-a3-templates"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer[all] (>=0.15.3,<0.16.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pyaml (>=25.1.0,<26.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "cookiecutter (>=2.6.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
